Fix default values to be null

+++ models/src/main/java/com/vimeo/networking2/LiveTime.kt
@@ -8,7 +8,7 @@ data class LiveTime(
     /**
      * The amount of time per event that the user is allowed to live stream.
      */
-    val eventMaximum: Int? = null,
+    val eventMaximum: Long? = null,
 
     /**
      * The amount of time this month, in seconds, that the user can live stream.
